Improve plugin management
This does a few things to help plugin configuration have fewer errors
when making changes and syncing across machines.

First, we defer loading our packer configuration in plugins.lua until we
explicitly run a sync command. This involves moving packer to an "opt"
plugin, disabling the default commands, and only sourcing the file and
running packer.sync when we run an explicit :Pack command. To facilite
intelligently reloading lua configs, we invalidate plugin files on
sourcing MYVIMRC. Definitely room for improvement here, but seems to
work fine for now.

Next, we bind the plugin config to the per-plugin "config" option. What
this buys us in particular is letting packer decide when it's
appropriate to load a plugin's config, meaning fewer errors for things
like attempting to load a configuration for a plugin that isn't
installed.

Also, I've noticed that because "sync" is really the only command I ever
run, it's the only command I've chosen to bind. Not hard to change if
that turns out to not be true.