Replace vim-go with vim-clover
With the addition of treesitter and LSP, we're down to exactly one
feature from vim-go that I still use: coverage. The downside is that
they keep adding more and more features that I have to explicitly
disable, and even then, unwanted plugin behavior leaks through.

While vim-closer is in no way a real replacement, it does exactly what I
need it to do and nothing else, at a quality that is TBD.