Go all in on vim-sleuth
This means not using a global editorconfig and expecting sleuth to
override it, since it seems to have a preference for editorconfigs over
sleuthing (which is the opposite of how it was configured before).
Unfortunately that previous config is no longer possible without pinning
sleuth.

There's a potential world in which we drop the idea of a global
editorconfig, or maybe just settings for any languages that don't have
universally agreed upon standards (e.g., JS) while keeping languages
that do have those standards (e.g., python). In favor of neovim's way of
handling indentation by filetype, which I have no idea how it works.

This might also be a good incentive to start dropping proper
editorconfigs in projects that I work on.