Load plugin config through packer
This makes for a grosser config and a necessity to re-sync (or at least
re-compile) packer whenever any plugin configuration changes.

But I think the benefit is that I basically never see errors on first
load or on making plugin changes, because there won't ever be a config
without having loaded the appropriate plugin.

We'll see how this pans out.