Pass bufnr more consistently
While I'm not sure if this ever manifested in practice, when we have a
specific buffer number, it's better to pass explicitly. This allows
these functions to be called from other buffers (which I'm not sure I
ever do), and prevents a race where a buffer is opened and switched away
from before an async registration would finish (which I'm not sure ever
happened).