Swap to neo-tree
The actual reason is I'm trying to switch to neotest, and the
combination of neotest and dirvish was giving me an autocmd error I was
not that interested in debugging.

The reason I wasn't that interested in debugging was I effectively tried
to use dirvish as "netrw without the bugs". It did a pretty good job,
but I spent a bizarre amount of effort trying to build a small
compatibility layer ontop of it. Dropping that effort for something more
out-of-the-box seems smart.

I tried mini.files and nvim-tree.lua, but neither of them seem
compatible with the oil-and-vinegar philosophy "split explorer", where
the buffer is taken over by the file browser. On that note I also looked
briefly at oil.nvim, which I could definitely see myself using, but I'm
not sure that the "edit your filesystem like a normal Neovim buffer"
philosophy is better.