Swap from neo-tree to oil
It took only a couple hours to run into a pretty critical bug for my
intended workflow: the neotest summary pane was being hijacked by
neo-tree. THe fact that there was already an issue conflicting with
Dirvish means that it's at least as likely to be a neotest issue as a
neo-tree issue, but it does seem a lot simpler than neo-tree, so we're
going to roll with oil for a minute and see if that doesn't make me
happy.