Swap to lualine from lightline
The short version of why I'm doing this is that it seems easy to add
support for the "recording" messages to show up in the status line if I
swap. There's probably a decent way to do it in lightline as well, but I
think my ability to configure things intuitively matters, too.