Swap from prettierd back to prettier
I keep seeing this issue where particularly long files get truncated
somewhere in the middle with format on save. Unfortunately it's not easy
to reproduce. No idea what the issue is, but running with prettier for a
while seems to have made the issue disappear completely.

If it never pops up again we have our answer.