Swap to mini.comment
After playing a bit with the built-in commenting from 0.10, I decided it
wasn't handling everything I wanted it to out of the box.

I probably could have continued to use commentary, but I guess for some
reason the plugin that the standard library functionality is based on
seems like a more "native" approach. If it annoys me I can always switch
back.