Change n_lines in mini.ai to 500
The default of 50 meant that it wouldn't respect long distances between
brackets, which with neovim plugin config was an absolute pain.

This might mean that the default `search_method = 'cover_or_next'` would
be better served as `cover`, although I'm going to try this out and see
if I get mad at it force jumping before I disable the ability to move to
the next match, which is inexplicably the same distance as searching for
matching brackets.