Add support for codeAction/resolve
They changed the spec such that a code action can be returned from the
server without the server actually having to compute the edits. Which
means that there is an edit to make, but we have to make a second
request.

The types here are less for safety and more for helping me remember what
on earth I was basing this logic on. Wouldn't be surprised if there are
more changes to be made in the future.