Configure multiple code actions on save
Add `source.fixAll`, although it's possible if some LSP I use takes a
long time to do that we swap it to `source.fixAll.biome`, because that's
what I'm really after today.

Also potentially fix an encoding issue that I've never seen in practice.
Be wary of that if I need to revert this someday.