Skip open API schema validation
The problem I'm running into is that it tries to use 3.1 validation in
all situations, and if I try to pin it to 3.0 validation, the YAML
language server complains that it doesn't support the particular draft
of JSON schema that is necessary for 3.0 anyway. So this basically stops
it from flashing an error message.