Simplify configuration with pipx/mise
The most important change here is I'm finally dropping the final git
submodule. The idea of git submodules are fine in theory, but every time
I try to do anything with them, it's such a pain that it's not worth it.
The alternative here is to make sure that pipx is installed, which is
easy and not a bad idea anyway, and then running dotbot through that.

The other change I'm trying out for a bit here is using mise to manage
the versions of selene and stylua. I'm not sure there's a benefit to
having them installed globally, and mason will manage them if I just
want them available in neovim. If I hate it, it's so very easy to
change.