Show more signs
I want the sign column to capture all signs EXCEPT for gitsigns. This is
a poor proxy, because it now reaches legacy signs (not in a namespace),
but not arbitrary signs, because I don't think I can negative match
against git signs, and I'm not familiar enough in the space for writing
a function to match, if that's even possible.