Only set foldexpr to use treesitter when a parser is available
For whatever reason, performance is abysmal on any large buffer where
this value is set and treesitter is not available. In particular,
opening an old toggle terminal back up that happened to have lots of
content on it.

This also has the side-benefit of keeping manual folding enabled on
non-treesitter buffers, which I've wanted once or twice.