Drop quicker
Neat concept, but I've run into more rough edges than I would like to at
this point. Most recently this is that treesitter syntax highlighting
breaks on multi-line nodes (most frequently, multi-line strings), and
that gO on the MANPAGER picks up a bunch of details that are not
particularly helpful in that view.