Unbind arrow keys in blink
It turns out I only ever use these for navigating history in command
mode. So if I'm pressing them, cancel and fall back.

The reason there's a function instead of just the string "cancel" is
because I want it to cancel AND fallback, whereas the string version
would stop after it successfully canceled.