Set winborder to rounded
This is still awkward in a couple of plug ins, primarily lazy.nvim and
mason, which both set a transparent background hover window and don't
explicitly set the border to "none".

But that's a super minor detail compared to getting borders on things
like native LSP hover and signature help. And I assume they'll fix it
before too long.