Set mise env vars before activation
So apparently the `conf.d/` directory is loaded before Homebrew's
installation of mise sets itself up, while `config.fish` is loaded
after.

In practice so far that's meaned if I open a new terminal tab in a
directory with a legacy idiomatic version file, I get a huge warning.
That said, there may be other unwanted effects, so this is generally
better even if that warning was gone.