Drop lazydev
As far as I can tell, the main feature of this is the lazy loading,
which for me is actually somewhat undesirable because it means that I
pay a short loading tax on every new file open rather than a longer
loading tax when the first lua file is opened. Which typically results
in a flash of warnings.

The reason for the duplication between `.luarc.jsonc` and `lua_ls.lua`
is that the former is scoped to this repo (and therefore needs the lazy
directory), and the latter is scoped to everything else.