Drop mason-lspconfig
The core idea with mason-lspconfig now is that it automatically enables
any server it has installed. What I want instead is to install anything
I have configured. Turns out, the mason registry's API has enough
information to do this, and mason-lspconfig doesn't really do anything
for me that I'm interested in.

Other than removing a dependency, this simplifies a few things
including:

- Making mason and lspconfig modules independent
- Performing all installation with mason-tool-installer
- Removing special cased configuration for languages that aren't on
  mason

I would still prefer to be able to get the list of configured servers
through something like `vim.lsp.get_enabled_servers` (which absolutely
does not exist), but for now, moving the list to its own file is not a
particularly big deal.