Properly include lazy deps in lua_ls config
I think based on how dynamic lazy.nvim is, some amount of scripting is
required to get all of the directories, unless I literally want to
enumerate every plug in I have installed in static config. While I could
re-add lazydev and get basically this for basically free, the problem is
lazydev does lazy-loading always, and that's annoying when you get a
flash of errors on every new lua file. Plus the amount of config it
actually takes to make this work seems to beat a full plug in by a
chunky margin.

This means I give up on my dream of being able to run
`lua-language-server --check` from the CLI, but honestly I wasn't
running it anyway considering how slow it was.