Request code actions on save in sequence
The original idea was I'd request all the code actions at once, then run
them in sequence. The problem is, at least for biome, those edits can
conflict with each other, meaning by the time one of them is applied,
the other one is stale and invalid.

This might add a small amount of latency, but it is probably the
simplest way to ensure that saving a second time doesn't make a second
edit.