Fix mise activation in nested shells
I was trying to avoid double activating mise, but (a) I don't think
that's necessary, and (b) if I open a nested shell such as a neovim
terminal, using an exported env var to detect activation is not actually
an accurate mechanism for doing that.

If this causes problems, next steps would be to read the script and
decide if there's another mechanism for detection. But hopefully that
isn't necessary.