Drop mason in favor of mise
It seems like every single tool I was installing with mason was also
available one way or another through mise. The ability to drop a package
manager seems nice, especially as I'm setting up a new device.

In theory it's not particularly different in terms of all of those tools
auto-installing before when I opened neovim for the first time, with the
exception of a small handful like go which weren't necessarily going to
be installed, but in practice, they were.

The one major change is the versions are checked into VCS, but outside
of maybe temporarily swapping to neovim nightly (which I've been moving
out of the habit of anyway), everything at "latest" is pretty much
right.